requirements.txt
faster-whisper==1.0.3
TTS==0.10.1
Flask==3.1.1
joblib==1.3.2
scikit-learn==1.3.2
sounddevice==0.5.2
soundfile==0.13.1
numpy==1.22.4
torch>=1.10.0
sentence-transformers==2.2.2
huggingface_hub<=0.15.1
rapidfuzz==3.9.6
PyYAML==6.0.2
webrtcvad==2.0.10
unicode